+++ public/admin/_app/immutable/nodes/24.CcMAT3pu.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/B_i1nuNu.js";export{e as component};